+++ pkg/kubelet/dockershim/security_context_test.go
@@ -25,7 +25,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	runtimeapi "k8s.io/kubernetes/pkg/kubelet/api/v1alpha1/runtime"
-	"k8s.io/kubernetes/pkg/securitycontext"
+	"k8s.io/kubernetes/pkg/kubelet/dockertools/securitycontext"
 )
 
 func TestModifyContainerConfig(t *testing.T) {
