+++ pkg/kubelet/dockertools/securitycontext/BUILD
@@ -13,21 +13,29 @@ go_library(
     srcs = [
         "doc.go",
         "fake.go",
-        "util.go",
+        "provider.go",
+        "types.go",
     ],
     tags = ["automanaged"],
     deps = [
-        "//pkg/api:go_default_library",
         "//pkg/api/v1:go_default_library",
+        "//pkg/kubelet/container:go_default_library",
+        "//pkg/kubelet/leaky:go_default_library",
+        "//pkg/securitycontext:go_default_library",
+        "//vendor:github.com/docker/engine-api/types/container",
     ],
 )
 
 go_test(
     name = "go_default_test",
-    srcs = ["util_test.go"],
+    srcs = ["provider_test.go"],
     library = ":go_default_library",
     tags = ["automanaged"],
-    deps = ["//pkg/api/v1:go_default_library"],
+    deps = [
+        "//pkg/api/testing:go_default_library",
+        "//pkg/api/v1:go_default_library",
+        "//vendor:github.com/docker/engine-api/types/container",
+    ],
 )
 
 filegroup(
