+++ pkg/securitycontext/fake.go
@@ -19,8 +19,6 @@ package securitycontext
 import (
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/v1"
-
-	dockercontainer "github.com/docker/engine-api/types/container"
 )
 
 // ValidSecurityContextWithContainerDefaults creates a valid security context provider based on
@@ -31,18 +29,6 @@ func ValidSecurityContextWithContainerDefaults() *v1.SecurityContext {
 		Capabilities: &v1.Capabilities{},
 		Privileged:   &priv,
 	}
-}
-
-// NewFakeSecurityContextProvider creates a new, no-op security context provider.
-func NewFakeSecurityContextProvider() SecurityContextProvider {
-	return FakeSecurityContextProvider{}
-}
-
-type FakeSecurityContextProvider struct{}
-
-func (p FakeSecurityContextProvider) ModifyContainerConfig(pod *v1.Pod, container *v1.Container, config *dockercontainer.Config) {
-}
-func (p FakeSecurityContextProvider) ModifyHostConfig(pod *v1.Pod, container *v1.Container, hostConfig *dockercontainer.HostConfig, supplementalGids []int64) {
 }
 
 // ValidInternalSecurityContextWithContainerDefaults creates a valid security context provider based on
